Permit an increased number of point sources
The original source-based July 2001 inventory appears to have an
incorrect spatial arrangement for gas species. In an effort to correct
this, the species-based July 2001 inventory gas species concentrations
were melded into the source-based July 2001 inventories. The
source-based July 2001 inventories have a different total number and
numbered location that corresponds to each point source, so this file
needed to be change to allow for a greater number of point sources.